[Refactor] Separate compilation error into it's own file
